requirements.txt
anyio==3.6.1
APScheduler==3.9.1
black==22.6.0
cachetools==5.2.0
certifi==2022.6.15
click==8.1.3
feedparser==6.0.10
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-telegram-bot==20.0a4
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
tomli==2.0.1
tornado==6.2
tzdata==2022.2
tzlocal==4.2
